Add convenience methods for common status queries
- get_battery_level(): returns SOC percentage 0-100%
- is_charging(): check if battery is currently charging
- is_battery_present(): check if battery is detected
- is_vbus_good(): check if USB/external power connected
- power_off(): trigger soft power off

Also fix examples to cast ChipVersion to u8 for defmt compatibility.

+++ examples/test_pmic_async.rs
@@ -74,7 +74,7 @@ async fn init_pmic(i2c: I2c<'_, Async>) -> Result<(), AxpError<I2cError>> {
     // Read chip ID using low-level API async
     let chip_id = axp.ll.chip_id().read_async().await?;
     info!("Chip ID high: {}, version: {}, low: {}",
-          chip_id.chip_id_high(), chip_id.chip_version(), chip_id.chip_id_low());
+          chip_id.chip_id_high(), chip_id.chip_version() as u8, chip_id.chip_id_low());
 
     // Read battery percentage (SoC) using low-level API
     let soc = axp.ll.battery_percentage().read_async().await?;
