+++ examples/test_pmic_blocking.rs
@@ -66,7 +66,7 @@ fn init_pmic(i2c: I2c<'_, Blocking>) -> Result<(), AxpError<I2cError>> {
     // Read chip ID using low-level API
     let chip_id = axp.ll.chip_id().read()?;
     info!("Chip ID high: {}, version: {}, low: {}",
-          chip_id.chip_id_high(), chip_id.chip_version(), chip_id.chip_id_low());
+          chip_id.chip_id_high(), chip_id.chip_version() as u8, chip_id.chip_id_low());
 
     // Read battery percentage (SoC) using low-level API
     let soc = axp.ll.battery_percentage().read()?;
